Add gulp integration in order to speed up compile time and better structure js files
The original version had all js files in a single file, making it
difficult to edit or find parts of the project.  Now the files are
broken out and gulp is used to bring them back together and then
minify.  This will be useful for any later integration of plugins as
well.

The checkin files which appear to be a seperate project entirely were
moved into the parent directory in order to prevent them from being
confused as part of this project.

Signed-off-by: Ethan Garrison <egarrison@varickmm.com>