Leaflet Updates
Updated old versions of angular-leaflet and leaflet, worked on handling
events better. Found that if leaflet queries for location even when
location service in chrome is disabled, causing a chrome alert to be
thrown. Feature was eliminated since it was unneccessary and manually
checking for permissions is not currently possible across browsers.

Signed-off-by: Briana Vecchione <briana.vecchione@gmail.com>